syslog-ng: make it possible to install a fixed version of syslog-ng
Copy of the comment in the install_syslog_ng_version.sh file:

The problem is that apt-get do not preserve the version number while
it is resolving metapackage dependencies. (Which is logical. If there
is any version restriction we should list them in the metapackage.)

example:
  We have a metapackage "foo", which lists "biz" and "baz" as
  dependencies. The command apt-get install foo=42, will install
  the latest greatest version of the biz and baz packages.

Since the dependencies of the syslog-ng metapackage can change from
version to version, (adding or removing modules) we can not hardcode
the package list in the apt-get command.

As a workaround this script queries the list of the dependencies of
the specific version of the syslog-ng metapackage. Parses the output,
than generates an apt-get install command where each package has the
same fixed version number in the form of <package-name>=VERSION.

Signed-off-by: Laszlo Szemere <laszlo.szemere@oneidentity.com>