dockerhub: add custom build hook
https://docs.docker.com/docker-hub/builds/advanced/

DockerHub provides some information about the state of the source
repository via environment variables. To access those environment
variables we have to override the default build command via build
hooks.

This is the original example from the documentation, only adding
the SYSLOG_NG_VERSION build argument.

The Dockerfile uses the SYSLOG_NG_VERSION to install a specific
version of the syslo-ng package. If the provided version is not
available in the OBS repository the build command will fail.

This way we can guarantee that the Docker image will contain the
right Syslog-ng, so we can Tag it accordingly.

Signed-off-by: Laszlo Szemere <laszlo.szemere@oneidentity.com>