syslog-ng: replace deprecated apt-key add to add apt repo keys
apt-key(8) will last be available in Debian 11 and Ubuntu 22.04. [1]

The key MUST be downloaded over a secure mechanism like HTTPS
to a location only writable by root, which SHOULD be /usr/share/keyrings.
The key MUST NOT be placed in /etc/apt/trusted.gpg.d or loaded by apt-key add. [2]

[1] https://manpages.debian.org/testing/apt/apt-key.8.en.html
[2] https://wiki.debian.org/DebianRepository/UseThirdParty#OpenPGP_Key_distribution

Signed-off-by: Gabor Nagy <gabor.nagy@oneidentity.com>